% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrs_geoms.R
\name{geom_line_nrs}
\alias{geom_line_nrs}
\alias{geom_col_nrs}
\alias{geom_text_nrs}
\title{Custom ggplot geoms with NRS defaults}
\usage{
geom_line_nrs(...)

geom_col_nrs(...)

geom_text_nrs(...)
}
\description{
These geoms are normal ggplot geoms. The only difference is they specify some
visual properties (such as size, width, colour).
}
