% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nrs_palette}
\alias{nrs_palette}
\title{Palette of colours used by National Records of Scotland for charts.}
\format{A list of named string vectors
\describe{
  Colours are specified as hex triplets
}}
\source{
\url{https://www.nrscotland.gov.uk/}
}
\usage{
nrs_palette
}
\description{
A palette of colours for making charts in the style used by National Records
of Scotland. Each topic area has a main colour with two variants (light and
dark). There is also a range of neutral colours (greys) which can be used for
any topic.
}
\keyword{datasets}
